Started adding the CurrentOrdersFragment

+++ waiter-core/src/main/java/com/example/waiterCore/domain/order/Order.kt
@@ -2,6 +2,6 @@ package com.example.waiterCore.domain.order
 
 data class Order (
     var tableId: Int,
-    var guestsCount: Long,
+    var guestsCount: Int,
     var orderItems: MutableSet<OrderItem> = mutableSetOf(),
 )
